requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
google-generativeai==0.8.3
google-genai>=0.2.0
httpx==0.26.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
Pillow==10.2.0
lxml==5.1.0
aiofiles==23.2.1
colorthief==0.2.1

# Supabase and Auth
supabase==2.7.4
postgrest==0.16.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Scheduling
apscheduler==3.10.4
